Added the ColorLogic class
It should let you take the screen output, sample it to find bricks, and then output the position of those bricks relative to the robot. Does not take into account the 3d size of the bricks yet though so it might not work perfectly. Also I had to guess at how the camera outputs color so the variable types might be wrong.